+++ src/main/java/Command/UtilityModule/MathCommand.java
@@ -29,7 +29,7 @@ public class MathCommand implements Command{
 
     public final static String HELP = "This command is for calculating math operations.\n"
                                     + "Command Usage: `" + Prefix.getDefaultPrefix() + "math` or `" + Prefix.getDefaultPrefix() + "calc` or `" + Prefix.getDefaultPrefix() + "m`\n"
-                                    + "Parameter: `-h | Math operation`\n"
+                                    + "Parameter: `-h | [Math Operation] | null`\n"
                                     + "Supported values: `pi, π, e, φ`\n";
     
     private static EmbedBuilder embed = new EmbedBuilder();