+++ src/main/java/Main/Main.java
@@ -117,6 +117,7 @@ public class Main {
         
         commands.put("say", new SayCommand());
         
+        commands.put("emoji", new EmojiCommand());
         commands.put("weather", new WeatherCommand());
         commands.put("w", new WeatherCommand());